+++ src/main/java/com/andreysotirov/website/controller/HomeController.java
@@ -1,0 +1,14 @@
+package com.andreysotirov.website.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+
+    @GetMapping("/home")
+    public String homePage() {
+        return "/home";
+    }
+
+}
