+++ src/main/java/com/andreysotirov/website/controller/LoginController.java
@@ -1,0 +1,14 @@
+package com.andreysotirov.website.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class LoginController {
+
+    @GetMapping("/login")
+    public String login() {
+        return "/login";
+    }
+
+}
